+++ example/elements/svg-fixed.js
@@ -1,10 +1,12 @@
+/* start example */
 import { svg } from '../../src'
 
 const fixedSvg = svg({
   height: 100,
   width: 100,
-  content: '\n  <example />\n',
+  content: '<example />',
   stretch: false,
 })
+/* end example */
 
 console.log(fixedSvg)