+++ example/elements/svg.js
@@ -1,9 +1,11 @@
+/* start example */
 import { svg } from '../../src'
 
 const stretchedSvg = svg({
   height: 100,
   width: 100,
-  content: '\n  <example />\n',
+  content: '<example />',
 })
+/* end example */
 
 console.log(stretchedSvg)