+++ src/index.js
@@ -1,25 +1,4 @@
-/**
- * Create a new element.
- * @param {MakeElementOptions} options Options to make a new element.
- * @param {string} options.name The name of the new element element.
- * @param {string} [options.content] The content to write inside of the element.
- * @param {object} [options.attributes] A map of attributes to add to the element.
- */
-export const makeElement = (options) => {
-  if (!options) throw new Error('Options were not passed.')
-  const {
-    name, content = '', attributes = {},
-  } = options
-  if (!name) throw new Error('Expected to see an element name.')
-  const attrs = Object.keys(attributes).reduce((acc, key) => {
-    const val = attributes[key]
-    if (val === undefined) return acc
-    return `${acc} ${key}="${val}"`
-  }, '')
-  const s = `${name}${attrs}`
-  const res = content ? `<${s}>${content}</${name}>` : `<${s}/>`
-  return res
-}
+import makeElement from './lib/make-element'
 
 /**
  * Draw a rounded corner by creating a C directive for the path element.
@@ -72,45 +51,12 @@ export const minify = (svg) => {
   return res
 }
 
-/**
- * Create an svg element.
- * @param {SVGOptions} options An option for creating an svg.
- * @param {number} options.width The width of the `svg`.
- * @param {number} options.height The height of the `svg`.
- * @param {string} options.content The content to put inside of the `svg`.
- * @param {boolean} [options.stretch=true] Expand the `svg` to the width of the container by not setting `width` and `height` attributes. Default `true`.
- */
-export const svg = (options) => {
-  const { width, height, stretch = true, content } = options
-  const res = makeElement({
-    name: 'svg',
-    attributes: {
-      version: '1.1',
-      xmlns: 'http://www.w3.org/2000/svg',
-      'xmlns:xlink': 'http://www.w3.org/1999/xlink',
-      viewBox: `0, 0, ${width}, ${height}`,
-      width: !stretch && width ? `${width}px` : undefined,
-      height: !stretch && height ? `${height}px` : undefined,
-    },
-    content,
-  })
-  return res
-}
-
 /* documentary types/index.xml */
 /**
- * @typedef {Object} SVGOptions An option for creating an svg.
- * @prop {number} width The width of the `svg`.
- * @prop {number} height The height of the `svg`.
- * @prop {string} content The content to put inside of the `svg`.
- * @prop {boolean} [stretch=true] Expand the `svg` to the width of the container by not setting `width` and `height` attributes. Default `true`.
- *
  * @typedef {Object} Coordinate A coordinate used for drawing.
  * @prop {number} x The `x` position of the coordinate.
  * @prop {number} y The `y` position of the coordinate.
- *
- * @typedef {Object} MakeElementOptions Options to make a new element.
- * @prop {string} name The name of the new element element.
- * @prop {string} [content] The content to write inside of the element.
- * @prop {object} [attributes] A map of attributes to add to the element.
  */
+
+export { makeElement }
+export { svg, rect } from './elements'