+++ test/spec/elements/svg.js
@@ -1,6 +1,6 @@
-import { equal, ok, throws } from 'zoroaster/assert'
+import { equal } from 'zoroaster/assert'
 import Context from '../../context'
-import { svg } from '../../../src'
+import { svg } from '../../../src/elements'
 
 /** @type {Object.<string, (c: Context)>} */
 const T = {
@@ -11,7 +11,10 @@ const T = {
       height: 100,
       width: 100,
     })
-    equal(res, `<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0, 0, 100, 100">${content}</svg>`)
+    equal(res, `<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"
+     viewBox="0, 0, 100, 100">
+  ${content}
+</svg>`)
   },
   'creates an svg element without stretch'({ content }) {
     const res = svg({
@@ -20,7 +23,10 @@ const T = {
       width: 100,
       stretch: false,
     })
-    equal(res, `<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0, 0, 100, 100" width="100px" height="100px">${content}</svg>`)
+    equal(res, `<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"
+     viewBox="0, 0, 100, 100" width="100px" height="100px">
+  ${content}
+</svg>`)
   },
 }