+++ test/spec/make-element.js
@@ -1,6 +1,6 @@
-import { equal, ok, throws } from 'zoroaster/assert'
+import { equal, throws } from 'zoroaster/assert'
 import Context from '../context'
-import { makeElement } from '../../src'
+import makeElement, { makeAttrs } from '../../src/lib/make-element'
 
 /** @type {Object.<string, (c: Context)>} */
 const T = {
@@ -8,32 +8,23 @@ const T = {
   async 'throws when no options are passed'() {
     await throws({
       fn: makeElement,
-      message: 'Options were not passed.',
-    })
-  },
-  async 'throws when no element is passed'() {
-    await throws({
-      fn: makeElement,
-      args: [{}],
-      message: 'Expected to see an element name.',
+      message: 'The element name was not passed.',
     })
   },
   'can create a new element'() {
-    const el = makeElement({
-      name: 'test',
-    })
+    const el = makeElement('test')
     equal(el, '<test/>')
   },
   'can create a new element with content'({ content }) {
-    const el = makeElement({
-      name: 'test',
+    const el = makeElement('test', {
       content,
     })
-    equal(el, `<test>${content}</test>`)
+    equal(el, `<test>
+  ${content}
+</test>`)
   },
   'can create a new element with attributes'() {
-    const el = makeElement({
-      name: 'test',
+    const el = makeElement('test', {
       attributes: {
         test: true,
         'font-color': 'green',
@@ -42,8 +33,7 @@ const T = {
     equal(el, '<test test="true" font-color="green"/>')
   },
   'ignores undefined attributes'() {
-    const el = makeElement({
-      name: 'test',
+    const el = makeElement('test', {
       attributes: {
         test: true,
         'font-color': 'green',
@@ -54,7 +44,7 @@ const T = {
     equal(el, '<test test="true" font-color="green"/>')
   },
   'can create a new element with attributes and content'({ content }) {
-    const el = makeElement({
+    const el = makeElement('test', {
       name: 'test',
       attributes: {
         test: true,
@@ -62,8 +52,59 @@ const T = {
       },
       content,
     })
-    equal(el, `<test test="true" font-color="green">${content}</test>`)
+    equal(el, `<test test="true" font-color="green">
+  ${content}
+</test>`)
+  },
+  'creates new element with array content'({ content }) {
+    const el = makeElement('test', {
+      content: [content, content],
+    })
+    equal(el, `<test>
+  ${content}
+  ${content}
+</test>`)
+  },
+  'creates new element with array content with new lines'({ content }) {
+    const t = makeElement('t', {
+      content: [content, content],
+    })
+    const el = makeElement('test', {
+      content: t,
+    })
+    equal(el, `<test>
+  <t>
+    ${content}
+    ${content}
+  </t>
+</test>`)
+  },
+  'filters out empty elements in content'({ content }) {
+    const el = makeElement('test', {
+      content: ['', content],
+    })
+    equal(el, `<test>
+  ${content}
+</test>`)
+  },
+}
+
+const MakeAttrs = {
+  'makes attributes with new lines'() {
+    const res = makeAttrs({
+      hello: 'world',
+      world: 'hello '.repeat(10).trim(),
+      test: 'test '.repeat(20).trim(),
+      test2: 'test2 '.repeat(20).trim(),
+      test3: 'test3 '.repeat(20).trim(),
+    }, 'el')
+    equal(res, ` hello="world" world="hello hello hello hello hello hello hello hello hello hello"
+    test="test test test test test test test test test test test test test test test test test test test test"
+    test2="test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2 test2"
+    test3="test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3 test3"`)
   },
 }
 
 export default T
+
+export { MakeAttrs }