+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 10 September 2018
+
+### 2.0.0
+
+- [doc] Add types as a section.
+- [feature] `makeElement`: change signature to accept name as the first argument, add attributes on new lines, content as array indent content (**breaking**).
+- [feature] Implement the `rect` element.
+
 ## 7 September 2018
 
 ### 1.0.0